Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 void main(){
-    int n;
-    scanf("%d",&n);
-    printf("Area: %d",3.14*n*n);
+    float n;
+    scanf("%f",&n);
+    printf("Area: %.2f",3.14*n*n);
 }
